fullstack/cashstatus/requirements.txt
asgiref==3.2.3
astroid==2.3.3
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
Django==3.0.2
django-rest-knox==4.1.0
djangorestframework==3.11.0
drf-yasg==1.17.0
flex==6.14.1
idna==2.8
inflection==0.3.1
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
jsonpointer==2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
packaging==20.0
pycparser==2.19
pylint==2.4.4
pyparsing==2.4.6
pytz==2019.3
PyYAML==5.3
requests==2.22.0
rfc3987==1.3.8
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
strict-rfc3339==0.7
typed-ast==1.4.0
uritemplate==3.0.1
urllib3==1.25.7
validate-email==1.3
wrapt==1.11.2
